% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidate_module}
\alias{candidate_module}
\title{candidate_module, candidate module}
\format{
An object of class \code{list} of length 4.
}
\usage{
candidate_module
}
\description{
candidate_module, Is the gene set from each local network by greedy algorithm,generated by `get_candidate_module`.
}
\keyword{datasets}
