% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{final_candidate_module}
\alias{final_candidate_module}
\title{final_candidate_module, Final candidate modules}
\format{
An object of class \code{list} of length 4.
}
\usage{
final_candidate_module
}
\description{
final_candidate_module, Is the final candidate module by intersecting modules from three protein networks in pairs,generated by `get_final_candidate_module`.
}
\keyword{datasets}
