% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_mut_survivalresult}
\alias{get_mut_survivalresult}
\title{Plot Kaplan-Meier survival curve.}
\usage{
get_mut_survivalresult(module, freq_matrix, sur)
}
\arguments{
\item{module}{The gene module,generated by `get_final_candidate_module`.}

\item{freq_matrix}{The mutations matrix,generated by `get_mut_status`.}

\item{sur}{A nx2 data frame of samples' survival data,the first line is samples' survival event and the second line is samples' overall survival.}
}
\value{
No return value
}
\description{
The function `get_mut_survivalresult` uses to draw the Kaplan-Meier survival curve based on the mutated status of candidate module.
}
\examples{
#load the data.
data(mut_status)
sur<-system.file("extdata","sur.csv",package="ProgModule")
sur<-read.delim(sur,sep=",",header=TRUE,row.names=1)
data(final_candidate_module)
#perform the function `get_mut_survivalresult`.
get_mut_survivalresult(module=final_candidate_module,freq_matrix=mut_status,sur)
}
