% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_network}
\alias{local_network}
\title{local_network, local network gene set}
\format{
An object of class \code{list} of length 4.
}
\usage{
local_network
}
\description{
local_network, the local network gene set of each gene by breadth-first algorithm,,generated by `get_local_network`.
}
\keyword{datasets}
