% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maf_data}
\alias{maf_data}
\title{maf_data, MAF file}
\format{
An object of class \code{data.frame} with 3745 rows and 10 columns.
}
\usage{
maf_data
}
\description{
maf_data, The patients' somatic mutation data, which in MAF format.
}
\keyword{datasets}
