% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MI}
\alias{MI}
\title{Calculate mutual information.}
\usage{
MI(mylist1, mylist2)
}
\arguments{
\item{mylist1}{Is input a list of samples' survival status.}

\item{mylist2}{Is input a list of samples' mutation status.}
}
\value{
The mutual information score
}
\description{
The function `MI` is used to calculate the mutual information score between samples' survival status and mutation status.
}
\examples{
\donttest{#load the data
data(mut_status)
sur<-system.file("extdata","sur.csv",package = "ProgModule")
sur<-read.delim(sur,sep=",",header=TRUE,row.names = 1)
#perform the function 'MI'
mut_matrix<-MI(mylist1 = as.numeric(mut_status[1,]),mylist2 = sur[,1])}
}
