% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{module}
\alias{module}
\title{module, gene set}
\format{
An object of class \code{character} of length 1.
}
\usage{
module
}
\description{
module, Original modular gene set.
}
\keyword{datasets}
