% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_status}
\alias{mut_status}
\title{mut_status, mutations matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 338 rows and 331 columns.
}
\usage{
mut_status
}
\description{
mut_status, the mutations matrix,generated by `get_mut_status`.
}
\keyword{datasets}
