% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net}
\alias{net}
\title{net, network}
\format{
An object of class \code{igraph} of length 76.
}
\usage{
net
}
\description{
net, Is a local network extracted from the ppi network.
}
\keyword{datasets}
