% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plotMutInteract_moduledata}
\alias{plotMutInteract_moduledata}
\title{plotMutInteract_moduledata}
\format{
An object of class \code{list} of length 7.
}
\usage{
plotMutInteract_moduledata
}
\description{
The data use for drawing mutually exclusive and co-occurrence plots.
}
\keyword{datasets}
