% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plotMutInteract_mutdata}
\alias{plotMutInteract_mutdata}
\title{plotMutInteract_mutdata}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 89 rows and 430 columns.
}
\usage{
plotMutInteract_mutdata
}
\description{
The data use for drawing mutually exclusive and co-occurrence plots.
}
\keyword{datasets}
