% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subnet}
\alias{subnet}
\title{subnet, network}
\format{
An object of class \code{igraph} of length 1624.
}
\usage{
subnet
}
\description{
subnet, Is a maximum connected subnet,extracted by mapping all genes to the ppi network.
}
\keyword{datasets}
