% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{univarCox_result}
\alias{univarCox_result}
\title{univarCox_result}
\format{
An object of class \code{numeric} of length 8103.
}
\usage{
univarCox_result
}
\description{
The result of Cox univariate analysis,generated by `get_univarCox_result`.
}
\keyword{datasets}
